refactor: move content from _init_.py to db.py
