+++ Backend/app/db/.env.example
@@ -1,1 +1,1 @@
-MONGO_URI = 
+MONGO_URI = your_mongodb_url
